+++ app/src/main/java/app/ddf/danskdatahistoriskforening/helper/VoiceRecorder.java
@@ -1,4 +1,4 @@
-package app.ddf.danskdatahistoriskforening;
+package app.ddf.danskdatahistoriskforening.helper;
 
 import android.media.MediaRecorder;
 import com.coremedia.iso.boxes.Container;
