+++ app/src/main/java/app/ddf/danskdatahistoriskforening/image/ImageviewerFragment.java
@@ -1,4 +1,4 @@
-package app.ddf.danskdatahistoriskforening;
+package app.ddf.danskdatahistoriskforening.image;
 
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
@@ -10,6 +10,8 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
+
+import app.ddf.danskdatahistoriskforening.R;
 
 public class ImageviewerFragment extends Fragment {
     private Uri imageUri;
