+++ app/src/main/java/app/ddf/danskdatahistoriskforening/item/ItemDescriptionFragment.java
@@ -10,11 +10,11 @@ import android.widget.EditText;
 import app.ddf.danskdatahistoriskforening.dal.Item;
 import app.ddf.danskdatahistoriskforening.R;
 
-public class RegisterDescriptionFragment extends Fragment {
+public class ItemDescriptionFragment extends Fragment {
     EditText itemDescription;
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        View layout = inflater.inflate(R.layout.fragment_register_description, container, false);
+        View layout = inflater.inflate(R.layout.fragment_item_description, container, false);
 
         itemDescription = (EditText) layout.findViewById(R.id.itemDescription);
