+++ app/src/main/java/app/ddf/danskdatahistoriskforening/main/ItemShowFragment.java
@@ -15,7 +15,7 @@ import app.ddf.danskdatahistoriskforening.Model;
 import app.ddf.danskdatahistoriskforening.R;
 
 
-public class ItemDetails extends Fragment {
+public class ItemShowFragment extends Fragment {
 
     private String detailsURI;
     private TextView itemheadlineView;
@@ -29,7 +29,7 @@ public class ItemDetails extends Fragment {
     private TextView postNummerView;
     protected Item currentItem;
 
-    public ItemDetails() {
+    public ItemShowFragment() {
         // Required empty public constructor
     }
 
@@ -41,7 +41,7 @@ public class ItemDetails extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        View layout = inflater.inflate(R.layout.fragment_item_details, container, false);
+        View layout = inflater.inflate(R.layout.fragment_item_show, container, false);
 
         itemheadlineView = (TextView) layout.findViewById(R.id.itemheadline);
         imageScrollView = (HorizontalScrollView) layout.findViewById(R.id.imageScrollView);
