+++ app/src/main/res/layout/activity_imageviewer.xml
@@ -47,7 +47,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="app.ddf.danskdatahistoriskforening.ImageviewerActivity">
+    tools:context="app.ddf.danskdatahistoriskforening.image.ImageviewerActivity">
 
 </RelativeLayout>
 -->
