+++ app/src/main/res/layout/activity_recording.xml
@@ -5,7 +5,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="app.ddf.danskdatahistoriskforening.RecordingActivity"
+    tools:context=".item.RecordingActivity"
     android:background="#000000">
 
     <ImageButton
