+++ app/src/main/res/layout/fragment_imageviewer.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="app.ddf.danskdatahistoriskforening.ImageviewerFragment">
+    tools:context=".image.ImageviewerFragment">
 
     <ImageView
         android:layout_width="fill_parent"
